fix(portability): keep address space in atomic pointers, widen flags to 32-bit
Two defects that made the decoupled-lookback and onesweep kernels unbuildable on
SPIR-V backends. Both were invisible on CUDA, and POCL was permissive enough to miss
them too -- they only surfaced on Intel Iris Xe via oneAPI.

1. _load_relaxed/_store_relaxed! did reinterpret(Ptr{U}, pointer(a, i)). pointer()
   returns Core.LLVMPtr{T,AS} with AS=1 (global); reinterpreting to Ptr{U} drops it to
   generic addrspace 0. SPIR-V allows casts only *to* generic, so this emitted an
   invalid module: "Casts from private/local/global address space are allowed only to
   generic". NVPTX tolerates the cast, which is why CUDA never complained. Reinterpret
   to Core.LLVMPtr{U,AS} instead, preserving the address space.

2. The lookback flag arrays were UInt8. SPIR-V/OpenCL define atomics only for 32- and
   64-bit types, so atomic load/store on a byte flag failed the Level Zero module build
   (ZE_RESULT_ERROR_MODULE_BUILD_FAILURE). Widened to UInt32, costing 3 bytes per block.

With both fixes, radix onesweep goes from failing on every type and size on Intel to
97/98 passing. Re-validated on CUDA: sort 352/352, decoupled lookback 560/560.

+++ src/sort/radix_sort.jl
@@ -539,7 +539,7 @@ end
         found = false
         b = iblock - 1
         while b >= 0
-            if _load_relaxed(flags, b * 256 + d + 1) == UInt8(1)
+            if _load_relaxed(flags, b * 256 + d + 1) == UInt32(1)
                 _decoupled_fence()                                     # value load after flag load
                 excl = _load_relaxed(incl, b * 256 + d + 1) + excl_partial
                 found = true
@@ -555,7 +555,7 @@ end
 
         _store_relaxed!(incl, iblock * 256 + d + 1, excl + local_count)   # publish inclusive prefix
         _decoupled_fence()                                                 # value before flag
-        _store_relaxed!(flags, iblock * 256 + d + 1, UInt8(1))
+        _store_relaxed!(flags, iblock * 256 + d + 1, UInt32(1))
         s_gbase[d + 1] = excl
         d += NI
     end
@@ -652,7 +652,7 @@ function _radix_sort!(
         digit_total = similar(v, UInt32, Int(_RS_SIZE))
         digit_base  = similar(v, UInt32, Int(_RS_SIZE))
         incl        = similar(v, UInt32, Int(_RS_SIZE) * num_blocks)
-        flags       = similar(v, UInt8,  Int(_RS_SIZE) * num_blocks)
+        flags       = similar(v, UInt32, Int(_RS_SIZE) * num_blocks)
     else
         # Single histogram buffer; no need to zero before each pass — the histogram kernel
         # zero-initializes its own shared-memory histogram and writes directly here.
@@ -722,7 +722,7 @@ function _radix_sort!(
             # digit_total accumulates via atomics and flags gates the look-back, so both start
             # zeroed each pass; incl is written before it is read.
             fill!(digit_total, UInt32(0))
-            fill!(flags, UInt8(0))
+            fill!(flags, UInt32(0))
             hist_os_kern!(hist, digit_total, p1, shift32, descending; ndrange)
             dbase_kern!(digit_base, digit_total; ndrange=(block_size,))
             os_kern!(p2, p1, hist, digit_base, incl, flags, shift32, descending; ndrange)
